requirements.txt
requests
kafka-python
#psycopg2-binary
pyodbc